WIP: Used a common ball display in both draggable and bouncing ball demos, also de-duped the alignment function for the popover menu
