+++ example/lib/demo_draggable_ball.dart
@@ -2,10 +2,10 @@ import 'package:example/infrastructure/ball_sandbox.dart';
 import 'package:flutter/material.dart';
 import 'package:follow_the_leader/follow_the_leader.dart';
 
-import 'ios_popover/_ios_popover_menu.dart';
+import 'ios_popover/ios_popover_menu.dart';
 
-/// Displays a blob that the user can drag, followed by a menu with an arrow
-/// that points in the direction of the blob.
+/// Displays a ball that the user can drag, followed by a menu with an arrow
+/// that points in the direction of the ball.
 class DraggableBallDemo extends StatefulWidget {
   const DraggableBallDemo({super.key});
 
@@ -15,8 +15,7 @@ class DraggableBallDemo extends StatefulWidget {
 
 class _DraggableBallDemoState extends State<DraggableBallDemo> {
   static const double _menuWidth = 100;
-  static const double _draggableBlogRadius = 50.0;
-  static const double _minimumDistanceFromEdge = 16;
+  static const double _draggableBallRadius = 50.0;
 
   final GlobalKey _screenBoundsKey = GlobalKey();
   final GlobalKey _leaderKey = GlobalKey();
@@ -52,44 +51,18 @@ class _DraggableBallDemoState extends State<DraggableBallDemo> {
       return;
     }
 
-    final focalPointInScreenBounds = _draggableOffset + const Offset(_draggableBlogRadius, _draggableBlogRadius);
+    final focalPointInScreenBounds = _draggableOffset + const Offset(_draggableBallRadius, _draggableBallRadius);
     final globalLeaderOffset = screenBoundsBox.localToGlobal(focalPointInScreenBounds);
 
     _globalMenuFocalPoint = globalLeaderOffset;
   }
 
-  /// Aligns the menu `Follower` with the `Leader` blob.
-  ///
-  /// We use this method in a `Follower.withDynamics` so that we can flip the menu
-  /// from one side of the blob to the other when we get close to the screen boundary.
   FollowerAlignment _alignMenu(Rect globalLeaderRect, Size followerSize) {
     final bounds = (_screenBoundsKey.currentContext?.findRenderObject() as RenderBox?)?.size ?? Size.zero;
-    print("Running aligner. Screen bound: $bounds");
 
-    late FollowerAlignment alignment;
-    if (globalLeaderRect.right + followerSize.width + _minimumDistanceFromEdge >= bounds.width) {
-      print(" - follower is too far to the right, switching to left");
-      // The follower hit the minimum distance. Invert the follower position.
-      alignment = const FollowerAlignment(
-        leaderAnchor: Alignment.centerLeft,
-        followerAnchor: Alignment.centerRight,
-        followerOffset: Offset(-20, 0),
-      );
-    } else if (globalLeaderRect.left - followerSize.width - _minimumDistanceFromEdge < 0) {
-      print(" - follower is too far to the left, switching to right");
-      // The follower hit the minimum distance. Invert the follower position.
-      alignment = const FollowerAlignment(
-        leaderAnchor: Alignment.centerRight,
-        followerAnchor: Alignment.centerLeft,
-        followerOffset: Offset(20, 0),
-      );
-    } else {
-      // We're not too far to the left or the right. Keep us wherever we were before.
-      alignment = _previousFollowerAlignment;
-    }
-
-    _previousFollowerAlignment = alignment;
-    return alignment;
+    final newAlignment = popoverMenuAligner(globalLeaderRect, followerSize, bounds, _previousFollowerAlignment);
+    _previousFollowerAlignment = newAlignment;
+    return newAlignment;
   }
 
   @override
@@ -100,11 +73,11 @@ class _DraggableBallDemoState extends State<DraggableBallDemo> {
           boundsKey: _screenBoundsKey,
           leaderKey: _leaderKey,
           followerKey: _followerKey,
-          blobOffset: _draggableOffset,
-          blobDecorator: (blob) {
+          ballOffset: _draggableOffset,
+          ballDecorator: (ball) {
             return GestureDetector(
               onPanUpdate: _onPanUpdate,
-              child: blob,
+              child: ball,
             );
           },
           followerAligner: _alignMenu,
