+++ example/lib/ios_popover/ios_toolbar.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 
-import '_ios_popover_menu.dart';
+import 'ios_popover_menu.dart';
 
 /// An iOS-style popover toolbar.
 ///
